+++ src/app/layout.tsx
@@ -1,8 +1,6 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
-import Header from '@/components/common/Header';
-import UserProvider from '@/context/UserContext';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -17,10 +15,8 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <UserProvider>
-      <html lang="ko">
-        <body className={inter.className}>{children}</body>
-      </html>
-    </UserProvider>
+    <html lang="ko">
+      <body className={inter.className}>{children}</body>
+    </html>
   );
 }
