+++ src/components/UserProfile.tsx
@@ -1,0 +1,30 @@
+'use client';
+
+import React, { useEffect } from 'react';
+import { useUserStore } from './store/userStore';
+import { createClient } from '@/supabase/client';
+
+const UserProfile = () => {
+  const { user, setUser } = useUserStore();
+  const supabase = createClient();
+
+  useEffect(() => {
+    async function getUser() {
+      const { data } = await supabase.auth.getUser();
+
+      if (data && data.user) {
+        setUser({
+          userId: data.user.id,
+          email: data.user.email
+        });
+      }
+      console.log(data);
+      console.log(user);
+    }
+    getUser();
+  }, [setUser]);
+
+  return <div></div>;
+};
+
+export default UserProfile;
