--- src/context/LoginProvider.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const LoginProvider = () => {
-  return <div>LoginProvider</div>;
-};
-
-export default LoginProvider;
